Add filler material for missing detail data

+++ src/org/starlo/climate/MainReport.java
@@ -4,7 +4,7 @@ public class MainReport
 {
     public Float temp;
     public Float pressure;
-    public Short humidity;
+    public Integer humidity;
     public Float temp_min;
     public Float temp_max;
 }
